Moving Summary Telemetry Types to the same file
